tests: update to use shared GHA workflows
